Clean up udev and Smack rules
Smack rules for device nodes should be applied in the udev package.
Security team should audit permissions for device nodes that are
needed to be changed that were being kept in this package (but again
should be done in the udev package as there were a broad grouping of
devices that may even conflict with default udev rules).

This change is likely a breaking change and will need corresponding
changes to other packages (including udev rules and manifest).

Change-Id: I509035ace21163e24231e825f44a9f96a988c47e
Signed-off-by: William Douglas <william.douglas@intel.com>